Gestión de errores de la BD

+++ src/main/java/com/example/accessingdatamysql/card/CardController.java
@@ -7,6 +7,7 @@ import com.example.accessingdatamysql.figure.Figure;
 import com.example.accessingdatamysql.figure.FigureService;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.server.ResponseStatusException;
 
 @Controller
 @RequestMapping(value = "/api")
@@ -29,7 +31,11 @@ public class CardController {
     @PostMapping(value = "/cards") // Map ONLY POST Requests
     public @ResponseBody Card addNewCard(@RequestBody Card card) {
         card.setFigures(new ArrayList<Figure>());
-        return this.cardService.saveCard(card);
+        try {
+            return this.cardService.saveCard(card);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+        }
     }
 
     @PostMapping(value = "/cards/{cardId}/figures/{figureId}")
@@ -37,7 +43,7 @@ public class CardController {
         Optional<Card> card = this.cardService.findCard(cardId);
         Optional<Figure> figure = this.figureService.findFigure(figureId);
         if (!card.isPresent()) {
-            return null;
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Card not found");
         }
         if (figure.isPresent()) {
 
@@ -50,8 +56,10 @@ public class CardController {
                 figure.get().getCards().add(card.get());
             if (!card.get().getFigures().contains(figure.get()))
                 card.get().getFigures().add(figure.get());
+
+            return this.cardService.saveCard(card.get());
         }
-        return this.cardService.saveCard(card.get());
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Figure not found");
     }
 
     @GetMapping(value = "/cards")
@@ -60,20 +68,27 @@ public class CardController {
     }
 
     @GetMapping(value = "/cards/{id}")
-    public @ResponseBody Optional<Card> getCardById(@PathVariable Long id) {
-        return this.cardService.findCard(id);
+    public @ResponseBody Card getCardById(@PathVariable Long id) {
+        Optional<Card> card = this.cardService.findCard(id);
+        if (card.isPresent())
+            return card.get();
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Card not found");
     }
 
     @DeleteMapping(value = "/cards/{id}")
-    public @ResponseBody String deleteCard(@PathVariable Long id) {
-        this.cardService.deleteCard(id);
-        return "Deleted";
+    public @ResponseBody void deleteCard(@PathVariable Long id) {
+        Optional<Card> card = this.cardService.findCard(id);
+        if (card.isPresent()) {
+            this.cardService.deleteCard(id);
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Card deleted");
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Card not found");
     }
 
     @DeleteMapping(value = "/cards")
-    public @ResponseBody String deleteAllCards() {
+    public @ResponseBody void deleteAllCards() {
         this.cardService.deleteAllCards();
-        return "Deleted all";
+        throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Cards deleted");
     }
 
     @DeleteMapping(value = "/cards/{cardId}/figures/{figureId}")
@@ -81,28 +96,35 @@ public class CardController {
         Optional<Card> card = this.cardService.findCard(cardId);
         Optional<Figure> figure = this.figureService.findFigure(figureId);
         if (!card.isPresent())
-            return "Card not found";
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Card not found");
         if (!figure.isPresent())
-            return "Figure not found";
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Figure not found");
 
         if (figure.get().getCards() == null) {
             figure.get().setCards(new ArrayList<Card>());
             if (card.get().getFigures() == null)
                 card.get().setFigures(new ArrayList<Figure>());
-            return "This card doesn't have figures";
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "This card has no figures");
         }
 
         card.get().getFigures().remove(figure.get());
         figure.get().getCards().remove(card.get());
-        return "Figure deleted from card";
+        throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Figure deleted from card");
     }
 
     @PutMapping(value = "/cards/{id}")
     public @ResponseBody Card updateCard(@RequestBody Card newCard, @PathVariable Long id) {
-        this.cardService.findCard(id).map(card -> {
+        Optional<Card> optCard = this.cardService.findCard(id);
+        if (!optCard.isPresent()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Card not found");
+        }
+        try {
+            Card card = optCard.get();
             card.setName(newCard.getName());
             return this.cardService.saveCard(card);
-        }).orElse(null);
-        return null;
+        } catch (Exception e) {
+            System.out.println(e.getCause());
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+        }
     }
 }
