+++ src/main/java/com/example/accessingdatamysql/game/GameController.java
@@ -11,6 +11,7 @@ import com.example.accessingdatamysql.user.Player;
 import com.example.accessingdatamysql.user.PlayerService;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,6 +21,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.server.ResponseStatusException;
 
 @Controller
 @RequestMapping(value = "/api")
@@ -35,21 +37,25 @@ public class GameController {
 
     @PostMapping(value = "/games") // Map ONLY POST Requests
     public @ResponseBody Game addNewGame(@RequestBody Game game) {
-        game.setMinigames(new ArrayList<Minigame>());
-        game.setPlayers(new ArrayList<Player>());
-        List<Player> players = game.getPlayers();
-        players.add(this.playerService.findPlayer(game.getCreator()).get());
-        game.setPlayers(players);
-        Player creator = this.playerService.findPlayer(game.getCreator()).get();
-        if (creator.getGamesPlayed() == null) {
-            creator.setGamesPlayed(new ArrayList<Game>());
-        }
-        List<Game> gamesPlayed = creator.getGamesPlayed();
-        game.setResults(new ArrayList<Result>());
-        gamesPlayed.add(game);
-        creator.setGamesPlayed(gamesPlayed);
+        try {
+            game.setMinigames(new ArrayList<Minigame>());
+            game.setPlayers(new ArrayList<Player>());
+            List<Player> players = game.getPlayers();
+            players.add(this.playerService.findPlayer(game.getCreator()).get());
+            game.setPlayers(players);
+            Player creator = this.playerService.findPlayer(game.getCreator()).get();
+            if (creator.getGamesPlayed() == null) {
+                creator.setGamesPlayed(new ArrayList<Game>());
+            }
+            List<Game> gamesPlayed = creator.getGamesPlayed();
+            game.setResults(new ArrayList<Result>());
+            gamesPlayed.add(game);
+            creator.setGamesPlayed(gamesPlayed);
 
-        return this.gameService.saveGame(game);
+            return this.gameService.saveGame(game);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+        }
     }
 
     @PostMapping(value = "/games/{gameId}/players/{playerId}")
@@ -57,7 +63,7 @@ public class GameController {
         Optional<Game> game = this.gameService.findGame(gameId);
         Optional<Player> player = this.playerService.findPlayer(playerId);
         if (!game.isPresent())
-            return null;
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
         if (player.isPresent()) {
             if (player.get().getGamesPlayed() == null)
                 player.get().setGamesPlayed(new ArrayList<Game>());
@@ -69,7 +75,7 @@ public class GameController {
             if (!game.get().getPlayers().contains(player.get()))
                 game.get().getPlayers().add(player.get());
         }
-        return this.gameService.saveGame(game.get());
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Player not found");
     }
 
     @PostMapping(value = "/games/{gameId}/minigames/{minigameId}")
@@ -77,7 +83,7 @@ public class GameController {
         Optional<Game> game = this.gameService.findGame(gameId);
         Optional<Minigame> minigame = this.minigameService.findMinigame(minigameId);
         if (!game.isPresent())
-            return null;
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
         if (minigame.isPresent()) {
             if (minigame.get().getGames() == null)
                 minigame.get().setGames(new ArrayList<Game>());
@@ -90,7 +96,7 @@ public class GameController {
                 game.get().getMinigames().add(minigame.get());
 
         }
-        return this.gameService.saveGame(game.get());
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Minigame not found");
     }
 
     @GetMapping(value = "/games")
@@ -99,51 +105,65 @@ public class GameController {
     }
 
     @GetMapping(value = "/games/{id}")
-    public @ResponseBody Optional<Game> getGameById(@PathVariable Long id) {
-        return this.gameService.findGame(id);
+    public @ResponseBody Game getGameById(@PathVariable Long id) {
+        Optional<Game> gOptional = this.gameService.findGame(id);
+        if (gOptional.isPresent())
+            return gOptional.get();
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
     }
 
     @GetMapping(value = "/games/{id}/players")
     public @ResponseBody List<Player> getPlayersByGame(@PathVariable Long id) {
-        return this.gameService.findGame(id).get().getPlayers();
+        Optional<Game> gOptional = this.gameService.findGame(id);
+        if (gOptional.isPresent())
+            return gOptional.get().getPlayers();
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
     }
 
     @GetMapping(value = "/games/names/{name}")
-    public @ResponseBody Optional<Game> getGameByName(@PathVariable String name) {
-        return this.gameService.findGameByName(name);
+    public @ResponseBody Game getGameByName(@PathVariable String name) {
+        Optional<Game> gOptional = this.gameService.findGameByName(name);
+        if (gOptional.isPresent())
+            return gOptional.get();
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
     }
 
     @DeleteMapping(value = "/games/{id}")
-    public @ResponseBody String deleteGame(@PathVariable Long id) {
-        this.gameService.deleteGame(id);
-        return "Deleted";
+    public @ResponseBody void deleteGame(@PathVariable Long id) {
+        Optional<Game> gOptional = this.gameService.findGame(id);
+        if (gOptional.isPresent()) {
+            this.gameService.deleteGame(id);
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Game deleted");
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
+
     }
 
     @DeleteMapping(value = "/games")
-    public @ResponseBody String deleteAllGames() {
+    public @ResponseBody void deleteAllGames() {
         this.gameService.deleteAllGames();
-        return "Deleted all";
+        throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Games deleted");
     }
 
     @DeleteMapping(value = "/games/{gameId}/players/{playerId}")
-    public @ResponseBody String deletePlayerFromGame(@PathVariable Long gameId, @PathVariable Long playerId) {
+    public @ResponseBody void deletePlayerFromGame(@PathVariable Long gameId, @PathVariable Long playerId) {
         Optional<Game> game = this.gameService.findGame(gameId);
         Optional<Player> player = this.playerService.findPlayer(playerId);
         if (!game.isPresent())
-            return "Game not found";
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
         if (!player.isPresent())
-            return "Player not found";
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Player not found");
 
         if (game.get().getPlayers() == null) {
             game.get().setPlayers(new ArrayList<Player>());
             if (player.get().getGamesPlayed() == null)
                 player.get().setGamesPlayed(new ArrayList<Game>());
-            return "This game doesn't have players";
+            throw new ResponseStatusException(HttpStatus.OK, "Game doesn't have players to remove");
         }
 
         game.get().getPlayers().remove(player.get());
         player.get().getGamesPlayed().remove(game.get());
-        return "Player deleted from game";
+        throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Player deleted from game");
 
     }
 
@@ -152,25 +172,29 @@ public class GameController {
         Optional<Game> game = this.gameService.findGame(gameId);
         Optional<Minigame> minigame = this.minigameService.findMinigame(minigameId);
         if (!game.isPresent())
-            return "Game not found";
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
         if (!minigame.isPresent())
-            return "Minigame not found";
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Minigame not found");
 
         if (game.get().getMinigames() == null) {
             game.get().setMinigames(new ArrayList<Minigame>());
             if (minigame.get().getGames() == null)
                 minigame.get().setGames(new ArrayList<Game>());
-            return "This game doesn't have minigames";
+            throw new ResponseStatusException(HttpStatus.OK, "Game doesn't have minigames to remove");
         }
 
         game.get().getMinigames().remove(minigame.get());
         minigame.get().getGames().remove(game.get());
-        return "Minigame deleted from game";
+        throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Minigame deleted from game");
     }
 
     @PutMapping(value = "/games/{id}")
     public @ResponseBody Game updateGame(@RequestBody Game newGame, @PathVariable Long id) {
-        this.gameService.findGame(id).map(game -> {
+        Optional<Game> gOptional = this.gameService.findGame(id);
+        if (!gOptional.isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Game not found");
+        try {
+            Game game = gOptional.get();
             game.setName(newGame.getName());
             game.setState(newGame.getState());
             game.setStartTime(newGame.getStartTime());
@@ -178,7 +202,8 @@ public class GameController {
             game.setCreator(newGame.getCreator());
             game.setWinner(newGame.getWinner());
             return this.gameService.saveGame(game);
-        }).orElse(null);
-        return null;
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+        }
     }
 }
